BUG: Fixed multi-value argument parsing
When a multi-value argument was passed to the command-line parser, all subsequent parameters were considered part of the argument.

The root cause of the problem was that the check that compared a parameter to a known argument was incorrect: the parameter name
contained the prefix (--argname) and this string was searched among argument names (such as argname).

Fixed by parsing each parameter after a multi-value argument. If a known argument is found, it is not added to the multi-value argument.